Java: InvocationContext API (#4790)
Roll prompt execution settings, kernel args, hooks, and return type into
an InvocationContext and add methods to Kernel and KernelFunction that
use an InvocationContext. The idea is that the invokeAsync methods that
take a KernelFunction, args and a return type are fine, but we don't
want to blow out the API with variations of invokeAsync for hooks, etc.
Also, prompt execution settings were in KernelArguments, which just
doesn't make sense. Now we can extend/add to the context as the need
arises without having to change the Kernel and KernelFunction API.
Sweet!