.Net: Improved Audio API by adding default setting values (#5026)
### Motivation and Context

<!-- Thank you for your contribution to the semantic-kernel repo!
Please help reviewers and future users, providing the following
information:
  1. Why is this change required?
  2. What problem does it solve?
  3. What scenario does it contribute to?
  4. If it fixes an open issue, please link to the issue here.
-->

Current audio interfaces have optional execution settings parameters in
order to be able to use audio API through Kernel in the future, where
execution settings are optional. But when using audio services directly,
this is not user friendly, because when optional execution settings are
not configured, services will throw an error.

In case of text-to-audio service, required field is voice. This PR sets
default value to `alloy` with the ability to overwrite.
For audio-to-text service, required field is filename. This PR sets
default value to `file.mp3` with the ability to overwrite.

With these changes execution settings are no longer required (they are
optional) for text-to-audio and audio-to-text conversion.

### Contribution Checklist

<!-- Before submitting this PR, please make sure: -->

- [x] The code builds clean without any errors or warnings
- [x] The PR follows the [SK Contribution
Guidelines](https://github.com/microsoft/semantic-kernel/blob/main/CONTRIBUTING.md)
and the [pre-submission formatting
script](https://github.com/microsoft/semantic-kernel/blob/main/CONTRIBUTING.md#development-scripts)
raises no violations
- [x] All unit tests pass, and I have added new tests where possible
- [x] I didn't break anyone :smile: