Add JsonProperty attributes for serialization
The changes include adding JsonProperty attributes to key properties in OpenAIModerationRequest.cs and OpenAIModerationResponse.cs files. The use of the JsonPropertyName attributes will ensure the correct serialization of properties. Also, it modifies the ModerationCategoryFlag and ModerationCategoryScore classes to use dictionary types instead of a list of custom classes.

Optimize DeserializeResponse method with optional parameter.