.Net: Fix serialization of OpenAIChatMessageContent type (#5301)
### Motivation and Context
This PR fixes the
https://github.com/microsoft/semantic-kernel/issues/5264 issue that
prevents auto calling of kernel functions created from prompt.

### Description  
1. The PR checks whether the function execution result type is
`ChatMessageContent` or its subtype and returns the
`ChatMessageContent.Content` property value if that's the case.
Otherwise, if the check was not provided, there would be an attempt to
serialize the class and because the `OpenAIChatMessageContent` is not
and can't be registered for polymorphic serialization on the
[KernelContent](https://github.com/microsoft/semantic-kernel/blob/d65d3b617b848d33cc1816b1756c3cdb65f4f337/dotnet/src/SemanticKernel.Abstractions/Contents/KernelContent.cs#L18)
class, the serialization would fail with the exception -
_"System.NotSupportedException: Runtime type
'Microsoft.SemanticKernel.Connectors.OpenAI.OpenAIChatMessageContent' is
not supported by polymorphic type
'Microsoft.SemanticKernel.KernelContent'."_ Alternatively, instead of
the type check above, the function result serializer could be configured
with a custom JsonTypeInfoResolver to handle polymorphic serialization
of the `OpenAIChatMessageContent` class. However, in this case, all the
public properties - ModelId, Metadata, Content, Items, Encoding, etc. of
the `OpenAIChatMessageContent` class would be unnecessarily serialized.
2. This PR also introduces the new
`ToolCallBehavior.FunctionResultSerializerOptions` to manage the
serialization behavior of function execution results. It can be used to
provide a JsonTypeInfoResolver for the polymorphic serialization of a
custom subclass of the `KernelContent` class.

<!-- Describe your changes, the overall approach, the underlying design.
These notes will help understanding how your code works. Thanks! -->

### Contribution Checklist

<!-- Before submitting this PR, please make sure: -->

- [x] The code builds clean without any errors or warnings
- [x] The PR follows the [SK Contribution
Guidelines](https://github.com/microsoft/semantic-kernel/blob/main/CONTRIBUTING.md)
and the [pre-submission formatting
script](https://github.com/microsoft/semantic-kernel/blob/main/CONTRIBUTING.md#development-scripts)
raises no violations
- [x] All unit tests pass, and I have added new tests where possible
- [x] I didn't break anyone :smile:

---------

Co-authored-by: Mark Wallace <127216156+markwallace-microsoft@users.noreply.github.com>