Respect Kernel plugin changes made during function auto-invocation
A KernelFunction can change when plug-ins are in the Kernel. Today, we build up the list of tools at the beginning of the auto-invocation loop but then don't update the list after a function has been invoked, which means those updates won't be reflected in subsequent calls made as part of the response to the LLM.  This PR updates the loop to regenerate the tool call information after each function invocation, in case they've been changed dynamically.