Update test assertions in GeminiPluginCollectionExtensionsTests
Changed the assertion methods used in GeminiPluginCollectionExtensionsTests. Instead of checking if the functions are the same, the test now asserts that the actual function is not null and that its name is equal to the expected function's name. This change makes the tests more precise and less prone to false positives.