.Net: Update test assertions in GeminiPluginCollectionExtensionsTests (#5549)
Changed the assertion methods used in
GeminiPluginCollectionExtensionsTests. Instead of checking if the
functions are the same, the test now asserts that the actual function is
not null and that its name is equal to the expected function's name.
This change makes the tests pass after cloning function was introduced
to kernel.