.Net: Add BERT ONNX embedding generation service (#5518)
Adds a new Microsoft.SemanticKernel.Connectors.Onnx component. As of
this PR, it contains one service,
BertOnnxTextEmbeddingGenerationService, for using BERT-based ONNX models
to generate embeddings. But in time we can add more ONNX-based
implementations for using local models.

This is in part based on
https://onnxruntime.ai/docs/tutorials/csharp/bert-nlp-csharp-console-app.html
and https://github.com/dotnet-smartcomponents/smartcomponents. It
doesn't support everything that's supported via sentence-transformers,
but we should be able to extend it as needed.

cc: @luisquintanilla, @SteveSandersonMS, @JakeRadMSFT