.Net: Upgrade projects from net6.0 to net8.0 and adjust LangVersion to 12 (#5819)
- Test projects targeting net6.0 are updated to net8.0
- LangVersion is consolidated to be 12 and set only in the top-level
Directory.Build.props
- RollForward was used to enable test projects targeting net6.0 to
execute in CI when there was only a net8.0 runtime present. These can
all be deleted.
- Microsoft.Net.Compilers.Toolset was used in a bunch of places to allow
for LangVersion 12 when on older toolsets. Now that the .NET 8 SDK is
used, this can all be deleted.
- Microsoft.CodeAnalysis.NetAnalyzers was pinned to the 8.0 version.
That can now be deleted.
- Allowing the analyzers to use a newer build by default flagged a bunch
of CA2007 violations to be addressed. These have all been fixed.