.Net: Don't limit [KernelFunction] to public methods (#6206)
A developer already need to opt-in a method on a plugin to being part of
the plugin by specifying the [KernelFunction] attribute; requiring that
method to also be public is superfluous, and means that a type's plugin
surface area must be a subset of its public surface area. That prohibits
patterns where a type wants to syntactically be a plugin but not expose
those APIs via its .NET public surface area.

(Curious to see if folks think this is controversial.)