.Net OpenAI SDK V2 - Phase 01 Embeddings + ClientCore (Feature Branch)  (#6898)
## ClientCore + Foundation

This PR is the first and starts the foundation structure and classes for
the V2 OpenAI Connector.

In this PR I also used the simpler `TextEmbeddings` service to wrap up a
vertical slice of the Service + Client + UT + IT + Dependencies needed
also to validate the proposed structure of folders and namespaces for
internal and public components.

### ClientCore

As part of this PR I'm also taking benefit of the `partial` keyword for
`ClientCore` class dividing its implementation per Service. In the
original V1 ClientCore the file was very big, and creating specific
files PR service/modality will make it simpler and easier to maintain.
  
## What Changed 

This change includes a new update from previous `Azure.Core` Pipeline
abstractions to the new `System.ClientModel` which is used by `OpenAI`
package.

Those include the update and addition of the below files:

- AddHeaderRequestPolicy - Adapted from previous
`AddHeaderRequestPolicy`
- ClientResultExceptionExtensions - Adapted from previous
`RequestExceptionExtensions`
- OpenAIClientCore - Merged with ClientCore (No more need for a
specialized Azure and OpenAI clients)
- ClientCore (Updated internals just with necessary for Text
Embeddings), merged `OpenAIClientCore` also into this one and made it
not as `abstract` class.
- OpenAITextEmbbedingGenerationService (Updated to use `ClientCore`
directly instead of `OpenAIClientCore`.

## Whats New 

- [PipelineSynchronousPolicy -
Azure.Core/src/Pipeline/HttpPipelineSynchronousPolicy.cs](https://github.com/Azure/azure-sdk-for-net/blob/8bd22837639d54acccc820e988747f8d28bbde4a/sdk/core/Azure.Core/src/Pipeline/HttpPipelineSynchronousPolicy.cs#L18)
This file didn't exist and was necessary to add as it is a dependency
for `AddHeaderRequestPolicy`
- Mockups added for `System.ClientModel` pipeline testing
- Unit Tests Covering
    - ClientCore
    - OpenAITextEmbeddingsGenerationService
    - AddHeadersRequestPolicy
    - PipelineSynchronousPolicy
    - ClientResultExceptionExtensions
- Integration Tests
    - OpenAITextEmbeddingsGenerationService (Moved from V1)

## What was Removed 

- OpenAIClientCore - This class was merged in ClientCore
- CustomHostPipelinePolicy - Removed as the new OpenAI SDK supports
Non-Default OpenAI endpoints.

## Unit & Integration Test

Differently from V1, this PR focus on individual UnitTest for the OpenAI
connector only.

With the target of above 80% code converage the Unit Tests targets
Services + Clients + Extensions & Utilities

The structure of folders and tested components on the UnitTests will
follow the same structure defined in project under test.