.Net: Migrate Azure Chat Completion Service to AzureOpenAI SDK v2 (#6984)
### Motivation and Context
This PR is the next step in a series of follow-up PRs to migrate
AzureOpenAIConnector to Azure AI SDK v2. It updates all code related to
AzureOpenAI ChatCompletionService to use the Azure AI SDK v2. One of the
goals of the PR is to update the code with a minimal number of changes
to make the code review as easy as possible, so almost all methods keep
their names as they were even though they might not be relevant anymore.
This will be fixed in one of the follow-up PRs.

### Description
This PR does the following:  
1. Migrates AzureOpenAIChatCompletionService, ClientCore, and other
model classes both use, to Azure AI SDK v2.
2. Updates ToolCallBehavior classes to return a list of functions and
function choice. This change is required because the new SDK model
requires both of those for the CompletionsOptions class creation and
does not allow setting them after the class is already created, as it
used to allow.
3. Adapts related unit tests to the API changes.

### Next steps
1. Add integration tests.  
2. Rename internal/private methods that were intentionally left with
old, irrelevant names to minimize the code review delta.

### Out of scope:
* https://github.com/microsoft/semantic-kernel/issues/6991

### Contribution Checklist

<!-- Before submitting this PR, please make sure: -->

- [x] The code builds clean without any errors or warnings
- [x] The PR follows the [SK Contribution
Guidelines](https://github.com/microsoft/semantic-kernel/blob/main/CONTRIBUTING.md)
and the [pre-submission formatting
script](https://github.com/microsoft/semantic-kernel/blob/main/CONTRIBUTING.md#development-scripts)
raises no violations
- [x] All unit tests pass, and I have added new tests where possible
- [x] I didn't break anyone :smile: