.Net: Copy AzureOpenAITextEmbeddingGenerationService to Connectors.AzureOpenAI project (#7022)
### Motivation and Context
This PR prepares the AzureOpenAITextEmbeddingGenerationService for
migration to the new Azure AI SDK v2. The
AzureOpenAITextEmbeddingGenerationService is copied to the
Connectors.AzureOpenAI project as is and excluded from compilation to
simplify code review of the next PR, which will refactor it to use the
new Azure SDK. The next PR will also add unit/integration tests, along
with service collection and kernel builder extension methods.

### Description
The AzureOpenAITextEmbeddingGenerationService class was copied as is to
the Connectors.AzureOpenAI project. The class build action was set to
none to exclude it temporarily from the compilation process until it
gets refactored to the new SDK.