.Net: Split ClientCore class (#7060)
### Motivation and Context
The ClientCore class has absorbed functionality relevant to different
Azure OpenAI-related services. As a first step to reduce its size, it
makes sense to split it into small chunks, where each chunk would be
related to a service that uses it. Later, we can move those chunks to
the services themselves.

### Description
This PR does the following:
1. Does not change any logic in any of the services.
2. Splits the `ClientCore` class into `ClientCore.ChatCompletion` and
`ClientCore.Embeddings` files.
3. Refactors the `AzureOpenAIChatCompletionService` and
`AzureOpenAITextEmbeddingGenerationService` to use the relevant
`ClientCore` pieces.
4. Removes the `AzureOpenAIClientCore` class