.Net: Tidying up AzureOpenAIChatCompletionService (#7073)
### Motivation, Context and Description
This PR fixes a small issue that would occur if the LLM started calling
tools that are not functions. Additionally, it renames the
`openAIClient` parameter of the `AzureOpenAIChatCompletionService` class
constructor to `azureOpenAIClient` to keep the name consistent with its
type. It also renames the
`AzureOpenAIChatMessageContent.GetOpenAIFunctionToolCalls` method to
`GetFunctionToolCalls` because the old one is not relevant anymore. The
last two changes are breaking changes and it will be decided in the
scope of the https://github.com/microsoft/semantic-kernel/issues/7053
issue whether to keep them or roll them back.