.Net: Copy OpenAITextToImageService related code to AzureOpenAI project (#7077)
### Motivation, Context and Description
This PR copies the `OpenAITextToImageService` class and related code,
including unit tests, from the `Connectors.OpenAIV2` project to the
`Connectors.AzureOpenAI` project. The copied classes have no functional
changes; they have only been renamed to have the `Azure` prefix and
placed into the corresponding
`Microsoft.SemanticKernel.Connectors.AzureOpenAI` namespace. All the
classes are temporarily excluded from the compilation process. This is
done to simplify the code review of follow-up PR(s) that will include
functional changes.

A few small fixes, unrelated to the main purpose of the PR, were made to
the XML documentation comments and logging in the
`OpenAIAudioToTextService` and `OpenAITextToImageService` classes.