.Net: Remove unnecessary azure chat message content classes (#7259)
### Motivation and Context
Two new classes, `AzureOpenAIChatMessageContent` and
`AzureOpenAIStreamingChatMessageContent`, were introduced during the
migration of AI connectors to the Azure.AI.OpenAI SDK v2 to speed up the
migration. However, it appeared that these classes are identical to
their non-Azure counterparts, `OpenAIChatMessageContent` and
`OpenAIStreamingChatMessageContent`. As a result, considering that they
do not provide any additional functionality over the OpenAI equivalents
at the moment, it was decided to drop them for now and use the OpenAI
ones instead. These classes might be added later if there is a need to
communicate Azure-specific details.

### Description
This PR removes the `AzureOpenAIChatMessageContent` and
`AzureOpenAIStreamingChatMessageContent` classes and refactors the
`ClientCore` to use the non-Azure equivalents.