.Net: Remove mapper type option and default redis prefixing to true. (#7356)
### Motivation and Context

As part of the new VectorStore pattern, we support passing custom models
to RecordCollection implementations.
There was an enum to specify which one you want to use, default or
custom, but we can just use the custom if provided
and fall back to the built in version if not.

Redis requires you to add a configurable prefix to record keys in order
for the record to be indexed by an index. Whether you want prefixing to
happen automatically is currently optional. Since records cannot be
indexed without it, and configuring the setting is painful with the new
vector store design, changing the default to true.

### Description

- Removing mapper type enums
- Switching redis prefix setting to true

### Contribution Checklist

<!-- Before submitting this PR, please make sure: -->

- [x] The code builds clean without any errors or warnings
- [x] The PR follows the [SK Contribution
Guidelines](https://github.com/microsoft/semantic-kernel/blob/main/CONTRIBUTING.md)
and the [pre-submission formatting
script](https://github.com/microsoft/semantic-kernel/blob/main/CONTRIBUTING.md#development-scripts)
raises no violations
- [x] All unit tests pass, and I have added new tests where possible
- [x] I didn't break anyone :smile: