.Net: RedisJson rename, adr rename and AzureAISearch integration tests disable (#7424)
### Motivation and Context

As part of the MemoryStore redesign making some changes to get ready to
merge into main.

Redis supports two modes when storing vectors. JSON and HashSets. I have
an implementation for HashSets but it doesn't have tests yet. I do
intend to add tests later and get it merged, but in the mean time, need
to make sure that the JSON implementation is named in a way that will
make it clear how it differs from the Hashset implementation.

### Description

Changes included:
1. Renaming the Redis record collection class to include JSON in the
name, since this implementation is specific to the JSON storage option
in Redis.
2. Renaming the ADR document to match the latest free number.
3. Disabling AzureAISearch integration tests, since they require infra
to be setup and isn't working as part of the build.


### Contribution Checklist

<!-- Before submitting this PR, please make sure: -->

- [x] The code builds clean without any errors or warnings
- [x] The PR follows the [SK Contribution
Guidelines](https://github.com/microsoft/semantic-kernel/blob/main/CONTRIBUTING.md)
and the [pre-submission formatting
script](https://github.com/microsoft/semantic-kernel/blob/main/CONTRIBUTING.md#development-scripts)
raises no violations
- [x] All unit tests pass, and I have added new tests where possible
- [x] I didn't break anyone :smile: