Python: removed underscore in .env for open_ai (#7625)
### Description

When running the getting-started Python notebooks, the user is
instructed to create and populate their own .env file. The values for
the OpenAI specifications included an underscore between 'open' and
'ai'. This led to an error when initializing the chat/text/embedding
model service.

I have removed the underscore from the base .env files, README, and in
the evaluate_with_prompt_flow.py script when reading from the .env file.