.Net: Function calling mvp model (#7974)
### Motivation and Context
This is the first PR that introduces an MVP model for the new function
calling abstraction.

The new function calling abstraction functionality has already been
implemented in this
https://github.com/microsoft/semantic-kernel/pull/6083/files. To
simplify the code review, the PR will be split into a series of small,
easy-to-review ones that will be merged into a separate
feature-function-calling-abstraction branch.

### Description
This PR introduces a few model classes that allow configuring function
calling behavior:
- `FunctionChoiceBehavior` - an abstract class that will be used as a
base class for all the choice behavior classes that will be added later
in separate PRs. It is intended to be used by AI connectors that will
call its `GetConfiguration` method to obtain configuration corresponding
to a provided choice behavior.
- `FunctionChoice` - a class that represents an LLM function choice.
- `FunctionChoiceBehaviorConfiguration` - a class that represents
function choice behavior configuration. Its instance is returned by the
`FunctionChoiceBehavior.GetConfiguration` method, which is called by AI
connectors to obtain configuration corresponding to a provided choice
behavior.
- `FunctionChoiceBehaviorConfigurationContext` - a class that represents
the configuration context that is supplied by AI connectors to the
`FunctionChoiceBehavior.GetConfiguration` method to provide context for
the configuration.

### Out Of Scope
- Deserialization
- Options
- Etc