.Net: Auto function choice behavior (#8067)
### Motivation, Context and Description
This PR adds a new `AutoFunctionChoiceBehavior` class that contains
information about functions to advertise to LLM, whether to auto-invoke
them or not, and the Auto choice instructing LLM to consider calling one
or many advertised functions. Additionally, it adds the new
`FunctionChoiceBehavior` property to the `PromptExecutionSettings`
class, which will be used by all AI connectors to configure function
choice behavior in a connector agnostic way.