.Net: Required function choice behavior (#8321)
### Motivation, Context and Description
This PR adds a new `RequiredFunctionChoiceBehavior` class that is used
by AI connectors to get a list of the functions to advertise to the AI
model and force it to call one or more functions. Additionally, it
provides information on whether the functions should be auto-invoked and
accepts a function selector callback to stop advertising functions if
the AI model prompt has already been answered.