Handle null parameter values in Handlebars
Passing a null parameter would previous cause a null reference exception for positional arguments.
It would throw a KernelException for hash arguments.
Change behavior of positional arguments to throw a KernelException to match behavior.

Adds tests for both cases as well as the positive test case where a argument is passed properly