.Net: Advertise required functions in the first request to AI model only (#8544)
### Motivation and Context
To fully finalize the new function-calling model, we need a mechanism to
prevent endless invocation of **required** functions by AI models. Until
this mechanism is in place, and to expedite the release of the new
function-calling model, it makes sense to align its behavior with that
of the current model. Specifically, required functions should be
advertised only in the initial request to the model and not in
subsequent requests, to prevent endless function calls by the model.

### Description
This PR modifies the `RequiredFunctionChoiceBehavior` class to advertise
functions only in the first request to the model and not in subsequent
requests of the same interaction with AI connector.

### Contribution Checklist

<!-- Before submitting this PR, please make sure: -->

- [x] The code builds clean without any errors or warnings
- [x] The PR follows the [SK Contribution
Guidelines](https://github.com/microsoft/semantic-kernel/blob/main/CONTRIBUTING.md)
and the [pre-submission formatting
script](https://github.com/microsoft/semantic-kernel/blob/main/CONTRIBUTING.md#development-scripts)
raises no violations
- [x] All unit tests pass, and I have added new tests where possible
- [x] I didn't break anyone :smile: