Python: Fix Anthropic Chat Completion. Fix agent chat broadcast message bug. (#8683)
### Motivation and Context

With the Anthropic AI Service, the chat completion path was missing some
functionality. To parse a chat history, there needs to exist an
overridden method that can form the chat history into the proper format
that Anthropic expects. An author role of SYSTEM is not allowed as a
role -- the SYSTEM prompt can exist as a property on the settings. If a
SYSTEM role is found in the chat history, and the system attribute on
the settings isn't defined then we can use it. We will always look for
the first system message in the chat history, if multiple are defined. A
required parameter, max_tokens, was also missing a default value. This
was added.

Additionally, an agent chat bug was discovered. During an agent chat, we
should only broadcast messages to other channels if they are different
from the current channel. In a scenario where two OpenAIChatCompletion
agents are conversing, they will be using the same channel. The bug made
it so that the message was broadcasted anyway and it caused duplicate
messages in the chat history.

<!-- Thank you for your contribution to the semantic-kernel repo!
Please help reviewers and future users, providing the following
information:
  1. Why is this change required?
  2. What problem does it solve?
  3. What scenario does it contribute to?
  4. If it fixes an open issue, please link to the issue here.
-->

### Description

This PR:
- fixes: #8639
- fixes the agent chat bug where messages were broadcasted to all
channels

<!-- Describe your changes, the overall approach, the underlying design.
These notes will help understanding how your code works. Thanks! -->

### Contribution Checklist

<!-- Before submitting this PR, please make sure: -->

- [X] The code builds clean without any errors or warnings
- [X] The PR follows the [SK Contribution
Guidelines](https://github.com/microsoft/semantic-kernel/blob/main/CONTRIBUTING.md)
and the [pre-submission formatting
script](https://github.com/microsoft/semantic-kernel/blob/main/CONTRIBUTING.md#development-scripts)
raises no violations
- [X] All unit tests pass, and I have added new tests where possible
- [X] I didn't break anyone :smile: