Move to using AsyncConnectionPool
This change moves to utilize psycopg's async connection pool, connection and cursor.
Changes were made dealing with how the connection pool is created.
A VectorStore itself is not a context manager, so managing closing the connection pool
was not possible without async tasks being cancelled incorrectly, even with the __del__ methods
that other VectorStores use.

Now, if a user uses a PostgresStore, they must pass in a connection pool they manage.
If they create a PostgresCollection without a connection pool, they have to use the
collection as a context manager which will close the connection pool on __aexit__.