.Net: Remove ReadOnlyMemoryConverter (#9014)
### Motivation, Context and Description
The ReadOnlyMemoryConverter was added as a stopgap to allow
ReadOnlyMemory<float> <-> float[] conversion because, at that time, the
conversion was not supported by .NET 6 and System.Text.Json v6.0.0.

Now, because SK uses System.Text.Json v8.0.0, there's no need for the
converter anymore. Moreover, it adds several unnecessary AOT-related
warnings when activating AOT for the Core package and a few other
packages as well.