.Net: Remove duplicating AOT attributes (#9045)
### Motivation, Context and Description
Previous PR to enable AOT for the Abstractions project/package annotated
several methods like `InvokeAsync` and `InvokeStreamingAsync` with the
`RequiresUnreferencedCode` and `RequiresDynamicCode` attributes. At the
same time, the same attributes were put on constructors of the classes
the methods belong to. This introduces unnecessary AOT warnings at the
methods' calling sites despite the fact that the classes were created
using AOT "friendly" constructors. This PR removes the attributes from
the methods, so no warnings are shown at their calling sites and keeps
the attributes on the classes' constructors to show the warnings at the
classes' construction sites.