.Net: Remove unnecessary DAM attributes (#9027)
### Motivation, Context and Description
Enum fields were never trimmed by the .NET trimmer, but the Enum
converter has the Type parameter annotated with the DAM attribute to
inform the calling code that it might be trimmed. Considering this,
there's no point in propagating the attributes up the calling stack, and
it makes sense to suppress that warning at the source. Moreover, in .NET
9, the Type constructor parameter is no longer annotated with the DAM
attribute.

This PR suppresses the "IL2067: UnrecognizedReflectionPattern..."
warning at the source and stops propagating the DAM attribute up the
calling stack. Additionally, it adds the
"Microsoft.SemanticKernel.Abstractions" assembly as TrimmerRootAssembly
to tell the trimmer to analyze every part of it.