.Net: {Azure}OpenAI Connectors Update to 2.1.0-beta.1 (#9078)
### Motivation and Context

Update to the latest pre-release of OpenAI and Azure OpenAI pre-release
packages following the new guidelines described in the ADR.

Moving forward connectors that target pre-releases will also be tagged
as pre-release as well as the `Microsoft.SemanticKernel` meta package
(that targets pre-release version of OpenAI and AzureOpenAI connectors.

THis update enable usage of the Realtime APIs provided by the underlying
SDKs allowing them to be used for breaking glass scenarios.