.NET: Processes support for sub-processes (#9095)
### Description

#### Enables use of sub-processes by adding a process to another process
as a step.

##### Event bubbling and visibility:
This PR adds the concept of visibility to events within a process. There
are two options, `Internal` which keeps the event within the process
it's fired in, and `Public` which allows the event to bubble out of the
process to parent processes or external systems. Events default to
`Internal`. This allows a parent process to receive events that
originate from within a sub-process.

##### Targets for process' external events:
Processes now expose targets for their external events. When a processes
defines a route for an external event by calling
`processBuilder.OnExternalEvent(...)...`, the target for this event can
now be retrieved by calling `process.GetTargetForExternalEvent(...)`.
This allows a parent process to route a step event to the specified
entry event of the sub-process.

#### Retrieve state of a running or completed process
The `*KernelProcessContext` now exposes a `GetStateAsync` method that
allows the state of the process to be retrieved on a started process.

Closes #9097 

### Contribution Checklist

<!-- Before submitting this PR, please make sure: -->

- [x] The code builds clean without any errors or warnings
- [x] The PR follows the [SK Contribution
Guidelines](https://github.com/microsoft/semantic-kernel/blob/main/CONTRIBUTING.md)
and the [pre-submission formatting
script](https://github.com/microsoft/semantic-kernel/blob/main/CONTRIBUTING.md#development-scripts)
raises no violations
- [x] All unit tests pass, and I have added new tests where possible
- [x] I didn't break anyone :smile:

---------

Co-authored-by: Ben Thomas <bentho@microsoft.com>