Add ability to close connection pool in mem store
This adds an __enter__ and __exit__ function to the PostgresMemoryStore
so that it can be used as a context manager.
Also modify the integration tests to utilize the context manager,
so that connections are freed up after the tests are run.
Without this change, the tests will hold onto the connections
and cause the database to run out of connections.