.Net: Make Core package AOT compatible (#9079)
### Motivation, Context and Description
This PR adds a new AOT-compatible API to the SK Core package and
refactors existing functionality to make it AOT-compatible.

99% of the changes involve new AOT-friendly overloads of methods to
create, add, or import plugins/functions that accept
JsonSerializerOptions (JSOs) as a parameter. JSOs are required for the
generation of plugins' functions schema, deserialization for functions'
parameters and return values marshaling, and serialization for telemetry
purposes. The remaining changes are related to making the vector store
functionality AOT-friendly and updating the test application to
reference the Core package for AOT compiler analysis.