.Net: Open SQLite connection when it's registered with connection string (#9407)
### Motivation and Context

<!-- Thank you for your contribution to the semantic-kernel repo!
Please help reviewers and future users, providing the following
information:
  1. Why is this change required?
  2. What problem does it solve?
  3. What scenario does it contribute to?
  4. If it fixes an open issue, please link to the issue here.
-->

Closes: https://github.com/microsoft/semantic-kernel/issues/9380

In DI methods, when SQLite connection string is not provided, we are
trying to get an instance from DI container, assuming that the
connection is already opened, and vector search extension is already
loaded. In other methods, where connection string is provided, we
initialize an instance of `SqliteConnection` on our side and load the
vector search extension. The missing part was to open a connection,
causing the following issue: `InvalidOperationException: ExecuteNonQuery
can only be called when the connection is open.`.

This PR contains a change to open a connection when we initialize an
instance of `SqliteConnection` on our side and also there are small
updates in XML documentation to better explain this behavior.

### Contribution Checklist

<!-- Before submitting this PR, please make sure: -->

- [x] The code builds clean without any errors or warnings
- [x] The PR follows the [SK Contribution
Guidelines](https://github.com/microsoft/semantic-kernel/blob/main/CONTRIBUTING.md)
and the [pre-submission formatting
script](https://github.com/microsoft/semantic-kernel/blob/main/CONTRIBUTING.md#development-scripts)
raises no violations
- [x] All unit tests pass, and I have added new tests where possible
- [x] I didn't break anyone :smile: