.Net: OpenAPI parameter resolution mechanism - part 1 (#9668)
### Motivation and Context
Today, the SK OpenAPI plugin's functionality does not support OpenAPI
operations that include parameters with the same name. For example, a
parameter named "id" can be used in the operation path, payload, header,
and query string, each representing the ID of a different entity in
those contexts. An attempt to import an OpenAPI operation containing
multiple parameters with the same name will fail with the error message:
"The function has two or more parameters with the same name '{parameter
name}'." Therefore, the SK OpenAPI plugin needs a way to gracefully
manage operations that include parameters with identical names.

### Description
This PR adds the `ArgumentName` property to the `RestApiPayloadProperty`
and `RestApiServerVariable` classes and renames the
`RestApiParameter.AlternativeName` property to
`RestApiParameter.ArgumentName` to better represent its purpose, as it
was used as an argument name before. The property allows associating
argument names with OpenAPI parameters and is used by the
`RestApiOperationRunner` to look up values for them in the arguments the
function was invoked with.
  
The next PR will allow consumers to access the OpenAPI document model
and provide argument names for the "clashing" parameters.

Additionally, this PR adds functionality that allows "freezing" the
classes to prevent their modification after the OpenAPI document has
been parsed and imported, and its operations have become available via
the `KernelFunction.Metadata.AdditionalProperties["operation"]` item.

Contributes to: https://github.com/microsoft/semantic-kernel/issues/4666