.Net: Prepare OpenAPI model collection properties to be modifiable. (#9735)
### Motivation, Context and Description
Currently, OpenAPI model classes use a read-only version of list and
dictionary types as types for collection properties. This is sufficient
for now, but it could become a breaking change later when they need to
be modifiable. In that case, they will have to be changed to
non-read-only types - specifically, IList<T> and IDictionary<TK, TV>.
This PR changes all the collection properties to be modifiable by
altering their types to IList<T> and IDictionary<TK, TV>.
   
~~The RestApiSecurityRequirement model class is not completely
freezeable at the moment and needs to be refactored to implement the
IDictionary interface instead of inheriting from Dictionary. This change
will allow for intercepting mutation calls and throwing an exception if
the class is frozen.~~

Contributes to: https://github.com/microsoft/semantic-kernel/issues/6884